uncertain loss: changed output type of getNoUncertain

+++ src/main/java/com/cs_pum/uncertain_mlc/losses/UncertainLoss.java
@@ -23,6 +23,6 @@ public interface UncertainLoss extends MultiLabelLossFunction, Measure {
      * Returns total count of uncertain labels
      * @return
      */
-    int getNoUncertain();
+    double getNoUncertain();
 
 }
